Refactored how core dependencies are initalized
Core dependencies loaded by ubash are no longer hard coded
as part of ubash code itself instead, they are retrieved by looking
the contents of "core" directory under $UPIPE_PATH. Also, it adds
the possibility of having a separated configuration for the core
dependencies. The configurations work in the same way as for the core
dependencies but, they are localized under $UPIPE_CONFIG_ROOT.
Each dependency can have their own configurations
when a configuration is found for the a dependency the environment
"<VERSIONEDNAME>_CONFIG_ROOT" and "<VERSIONEDNAME>_CONFIG_VERSION"
are exported automatically during ubash initialization. Also,
configurations can have "preInit" and "postInit" callbacks which
are executed automatically in between the initialization of the
core dependency.
The order from which the dependencies are loaded is determined by the
"depends" information inside of the "info.json". Therefore, if
you have a dependency that depends from a specific one. For instance,
ulauncher depends from uver, you can define this relationship by
adding the entry 'depends: "uver"' inside of the info.json in
ulauncher.